pyproject.toml
[tool.poetry]
name = "repl_python3_spookiest-time-of-the-year"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.8"
blend-modes = "^2.1.0"
Pillow = "^8.3.2"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
